Add stream configuration importing
The patch adds the ability to load streams from a user-imported
module.  In order to start dmon the user will have to add the path to
their module to PYTHONPATH and specify the module name with the
special environment variable, DMON_STREAM_MODULE.  This design is used
in Django and shouldn't be too hard to pick up.

A stream module simply binds some co-routines to module-level names
with the word "stream" in them.

There are also some minor fixes here and there.